pyproject.toml
[project]
name = "ff-data-analytics"
version = "0.1.0"
description = "Adavanced data architecture and analytics for Fantasy Football"
readme = "README.md"
requires-python = ">=3.13.6"
dependencies = [
    "google-api-python-client>=2.183.0",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "nflreadpy>=0.1.3",
    "oauth2client>=4.1.3",
    "pandas>=2.3.2",
    "polars>=1.33.1",
    "pyarrow>=21.0.0",
    "pydrive2>=1.21.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["ingest", "tools"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mdformat]
wrap = 100
number = false

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
